backend/requirements.txt
Django==5.0.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
django-filter==23.5
drf-yasg==1.21.7
python-decouple==3.8
python-dateutil==2.8.2
pytz==2024.1
uritemplate==4.1.1
inflection==0.5.1
